# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Multi-Agent Hybrid RAG API
  version: 0.1.0
  description: |
    Initial Phase 0 HTTP contract. All `/v1` operations require a verified
    Supabase JWT and an active workspace unless explicitly marked public.
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://{apiHost}
    description: Hosted API
    variables:
      apiHost:
        default: example.onrender.com
security:
  - bearerAuth: []
...
paths:
...
  /v1/auth/me:
    get:
      tags: [Authentication]
      operationId: getCurrentUser
      summary: Verify the Supabase access token and return its user identity
      responses:
        "200":
          description: Verified authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthenticatedUser"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/workspaces/{workspaceId}/access:
    get:
      tags: [Authentication]
      operationId: getWorkspaceAccess
      summary: Resolve the caller's RLS-protected workspace role
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Workspace membership and role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceAccess"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "503":
          $ref: "#/components/responses/Unavailable"
...
  /v1/retrieval/search:
    post:
      tags: [Retrieval]
      operationId: hybridSearch
      summary: Search indexed workspace documents using dense and keyword fusion
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetrievalRequest"
      responses:
        "200":
          description: Deterministically ranked workspace-scoped chunks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetrievalResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/BadRequest"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/conversations:
    get:
      tags: [Conversations]
      operationId: listConversations
      summary: List the current user's conversations
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Paginated conversations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationPage"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [Conversations]
      operationId: createConversation
      summary: Create a conversation
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationRequest"
      responses:
        "201":
          description: Conversation was created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"
  /v1/conversations/{conversationId}:
    get:
      tags: [Conversations]
      operationId: getConversation
      summary: Get a conversation and recent messages
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ConversationId"
      responses:
        "200":
          description: Conversation detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConversationDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/conversations/{conversationId}/messages:
    post:
      tags: [Conversations, Runs]
      operationId: createMessageRun
      summary: Create a user message and start a RAG or agent run
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/ConversationId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageRequest"
      responses:
        "202":
          description: Run was accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunAccepted"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
  /v1/runs/{runId}:
    get:
      tags: [Runs]
      operationId: getRun
      summary: Get the durable run snapshot
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/runs/{runId}/events:
    get:
      tags: [Runs]
      operationId: streamRunEvents
      summary: Stream live answer and run events
      description: |
        Authenticated SSE stream. Clients may send `Last-Event-ID` when
        reconnecting. Durable state is always available from the run endpoint.
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/RunId"
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Server-Sent Events using the schemas in events.asyncapi.yaml
          content:
            text/event-stream:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/runs/{runId}/cancel:
    post:
      tags: [Runs]
      operationId: cancelRun
      summary: Request cancellation of a non-terminal run
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - $ref: "#/components/parameters/RunId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Cancellation was accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OperationAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    WorkspaceId:
      name: X-Workspace-ID
      in: header
      required: true
      description: Active workspace. Membership is verified server-side.
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        maxLength: 512
    PageSize:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RunId:
      name: runId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Request validation or business-rule failure
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unauthorized:
      description: Missing, invalid, or expired authentication
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Forbidden:
      description: Authenticated actor lacks permission
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFound:
      description: Resource does not exist or is not visible to this actor
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Conflict:
      description: State or idempotency conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    RateLimited:
      description: Application or provider quota is exhausted
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds before retry, when known
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    Unavailable:
      description: Required provider is temporarily unavailable
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
...
    AuthenticatedUser:
      type: object
      additionalProperties: false
      required: [id, role]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: [string, "null"]
          format: email
        role:
          const: authenticated
    WorkspaceAccess:
      type: object
      additionalProperties: false
      required: [workspace_id, user_id, role]
      properties:
        workspace_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, reviewer, member]
    Problem:
      type: object
      additionalProperties: false
      required: [type, title, status, code, request_id]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        code:
          type: string
          pattern: "^[A-Z][A-Z0-9_]+$"
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
        request_id:
          type: string
          format: uuid
        retryable:
          type: boolean
          default: false
        fields:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
...
    RetrievalMode:
      type: string
      enum: [hybrid, dense, sparse]
    RetrievalFilters:
      type: object
      additionalProperties: false
      properties:
        document_ids:
          type: [array, "null"]
          maxItems: 20
          items:
            type: string
            format: uuid
        created_after:
          type: [string, "null"]
          format: date-time
        created_before:
          type: [string, "null"]
          format: date-time
        content_types:
          type: [array, "null"]
          maxItems: 4
          items:
            type: string
            enum: [application/pdf, text/plain, text/markdown, text/html]
        tags:
          type: [array, "null"]
          maxItems: 20
          items:
            type: string
            maxLength: 50
    RetrievalRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 2000
        mode:
          $ref: "#/components/schemas/RetrievalMode"
        limit:
          type: integer
          minimum: 1
          maximum: 20
          default: 6
        candidate_count:
          type: integer
          minimum: 1
          maximum: 100
          default: 30
        filters:
          $ref: "#/components/schemas/RetrievalFilters"
    RetrievalResult:
      type: object
      additionalProperties: false
      required:
        - chunk_id
        - document_id
        - chunk_index
        - content
        - page_start
        - page_end
        - filename
        - content_type
        - tags
        - document_created_at
        - rrf_score
        - final_rank
      properties:
        chunk_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        chunk_index:
          type: integer
        content:
          type: string
        page_start:
          type: integer
        page_end:
          type: integer
        section_heading:
          type: [string, "null"]
        char_start:
          type: [integer, "null"]
        char_end:
          type: [integer, "null"]
        token_count:
          type: [integer, "null"]
        filename:
          type: string
        title:
          type: [string, "null"]
        content_type:
          type: string
        tags:
          type: array
          items:
            type: string
        document_created_at:
          type: string
          format: date-time
        semantic_rank:
          type: [integer, "null"]
        sparse_rank:
          type: [integer, "null"]
        semantic_score:
          type: [number, "null"]
        sparse_score:
          type: [number, "null"]
        rrf_score:
          type: number
        final_rank:
          type: integer
    RetrievalResponse:
      type: object
      additionalProperties: false
      required:
        - trace_id
        - cache_hit
        - embedding_cache_hit
        - index_fingerprint
        - dense_candidate_count
        - sparse_candidate_count
        - embedding_ms
        - database_ms
        - total_ms
        - items
      properties:
        trace_id:
          type: string
          format: uuid
        cache_hit:
          type: boolean
        embedding_cache_hit:
          type: boolean
        index_fingerprint:
          type: string
          pattern: "^[a-f0-9]{32}$"
        dense_candidate_count:
          type: integer
          minimum: 0
        sparse_candidate_count:
          type: integer
          minimum: 0
        embedding_ms:
          type: number
          minimum: 0
        database_ms:
          type: number
          minimum: 0
        total_ms:
          type: number
          minimum: 0
        items:
          type: array
          items:
            $ref: "#/components/schemas/RetrievalResult"
...
    CreateConversationRequest:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
          maxLength: 200
    Conversation:
      type: object
      required: [id, workspace_id, owner_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        workspace_id:
          type: string
          format: uuid
        owner_id:
          type: string
          format: uuid
        title:
          type: [string, "null"]
        summary:
          type: [string, "null"]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ConversationPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Conversation"
        next_cursor:
          type: [string, "null"]
    MessageRole:
      type: string
      enum: [user, assistant, system]
    Message:
      type: object
      additionalProperties: false
      required: [id, conversation_id, role, content, created_at]
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        role:
          $ref: "#/components/schemas/MessageRole"
        content:
          type: string
        answer_status:
          type: [string, "null"]
          enum: [grounded, insufficient_evidence, failed, null]
        confidence:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        citations:
          type: array
          items:
            $ref: "#/components/schemas/Citation"
        created_at:
          type: string
          format: date-time
    ConversationDetail:
      allOf:
        - $ref: "#/components/schemas/Conversation"
        - type: object
          required: [messages]
          properties:
            messages:
              type: array
              items:
                $ref: "#/components/schemas/Message"
    CreateMessageRequest:
      type: object
      additionalProperties: false
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
          maxLength: 12000
        document_ids:
          type: array
          maxItems: 100
          items:
            type: string
            format: uuid
        force_mode:
          type: string
          enum: [auto, simple, agentic]
          default: auto
    RunMode:
      type: string
      enum: [simple, agentic]
    RunStatus:
      type: string
      enum:
        - accepted
        - running
        - awaiting_approval
        - cancelling
        - cancelled
        - completed
        - failed
        - timed_out
    RunAccepted:
      type: object
      additionalProperties: false
      required: [run_id, message_id, status, events_url]
      properties:
        run_id:
          type: string
          format: uuid
        message_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RunStatus"
        events_url:
          type: string
          format: uri-reference
    Run:
      type: object
      additionalProperties: false
      required:
        - id
        - conversation_id
        - status
        - mode
        - step_count
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
        conversation_id:
          type: string
          format: uuid
        status:
          $ref: "#/components/schemas/RunStatus"
        mode:
          $ref: "#/components/schemas/RunMode"
        current_node:
          type: [string, "null"]
        step_count:
          type: integer
          minimum: 0
          maximum: 8
        confidence:
          type: [number, "null"]
          minimum: 0
          maximum: 1
        answer_status:
          type: [string, "null"]
          enum: [grounded, insufficient_evidence, failed, null]
        output_message_id:
          type: [string, "null"]
          format: uuid
        approval_id:
          type: [string, "null"]
          format: uuid
        error:
          oneOf:
            - $ref: "#/components/schemas/Problem"
            - type: "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: [string, "null"]
          format: date-time
...
    Citation:
      type: object
      additionalProperties: false
      required: [citation_id, document_id, chunk_id, label, quote]
      properties:
        citation_id:
          type: string
          pattern: "^C[1-9][0-9]*$"
        document_id:
          type: string
          format: uuid
        chunk_id:
          type: string
          format: uuid
        label:
          type: string
        page:
          type: [integer, "null"]
          minimum: 1
        section:
          type: [string, "null"]
        quote:
          type: string
          maxLength: 1000
        source_url:
          type: string
          format: uri-reference
...
    OperationAccepted:
      type: object
      additionalProperties: false
      required: [operation_id, status]
      properties:
        operation_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [accepted]